# workspaces/backend/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage notebooks in a Kubernetes cluster.\nFor more information, visit https://www.kubeflow.org/docs/components/notebooks/",
        "title": "Kubeflow Notebooks API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/storageclasses": {
            "get": {
                "description": "Returns a list of all storage classes in the cluster.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storageclasses"
                ],
                "summary": "List storage classes",
                "operationId": "listStorageClasses",
                "responses": {
                    "200": {
                        "description": "Successful storage classes response",
                        "schema": {
                            "$ref": "#/definitions/api.StorageClassListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/workspacekinds": {
            "get": {
                "description": "Returns a list of all workspace kinds in the cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspacekinds"
                ],
                "summary": "List workspace kinds",
                "operationId": "listWorkspaceKinds",
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns a list of all available workspace kinds.",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceKindListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to list workspace kinds.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new workspace kind.",
                "consumes": [
                    "application/yaml"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspacekinds"
                ],
                "summary": "Create workspace kind",
                "operationId": "createWorkspaceKind",
                "parameters": [
                    {
                        "description": "Kubernetes YAML manifest of a WorkspaceKind",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "WorkspaceKind created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceKindEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to create WorkspaceKind.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict. WorkspaceKind with the same name already exists.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large. The request body is too large.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type. Content-Type header is not correct.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/workspacekinds/{name}": {
            "get": {
                "description": "Returns details of a specific workspace kind identified by its name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspacekinds"
                ],
                "summary": "Get workspace kind",
                "operationId": "getWorkspaceKind",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "jupyterlab",
                        "description": "Name of the workspace kind",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns the requested workspace kind details.",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceKindEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request. Invalid workspace kind name format.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to access the workspace kind.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found. Workspace kind does not exist.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ConflictError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "A human-readable description of the cause of the error.\nThis field may be presented as-is to a reader.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin indicates where the conflict error originated.\nIf value is empty, the origin is unknown.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCauseOrigin"
                        }
                    ]
                }
            }
        },
        "api.ErrorCause": {
            "type": "object",
            "properties": {
                "conflict_cause": {
                    "description": "ConflictCauses contains details about conflict errors that caused the request to fail.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ConflictError"
                    }
                },
                "validation_errors": {
                    "description": "ValidationErrors contains details about validation errors that caused the request to fail.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ValidationError"
                    }
                }
            }
        },
        "api.ErrorCauseOrigin": {
            "type": "string",
            "enum": [
                "INTERNAL",
                "KUBERNETES"
            ],
            "x-enum-varnames": [
                "OriginInternal",
                "OriginKubernetes"
            ]
        },
        "api.ErrorEnvelope": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.HTTPError"
                }
            }
        },
        "api.HTTPError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "cause": {
                    "description": "Cause contains detailed information about the cause of the error.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCause"
                        }
                    ]
                },
                "code": {
                    "description": "Code is a string representation of the HTTP status code.",
                    "type": "string"
                },
                "message": {
                    "description": "Message is a human-readable description of the error.",
                    "type": "string"
                }
            }
        },
...
        "api.StorageClassListEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storageclasses.StorageClassListItem"
                    }
                }
            }
        },
        "api.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "The field of the resource that has caused this error, as named by its JSON serialization.\nMay include dot and postfix notation for nested attributes.\nArrays are zero-indexed.\nFields may appear more than once in an array of causes due to fields having multiple errors.\n\nExamples:\n  \"name\" - the field \"name\" on the current resource\n  \"items[0].name\" - the field \"name\" on the first array entry in \"items\"",
                    "type": "string"
                },
                "message": {
                    "description": "A human-readable description of the cause of the error.\nThis field may be presented as-is to a reader.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin indicates where the validation error originated.\nIf value is empty, the origin is unknown.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCauseOrigin"
                        }
                    ]
                },
                "type": {
                    "description": "A machine-readable description of the cause of the error.\nIf value is empty, there is no information available.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/field.ErrorType"
                        }
                    ]
                }
            }
        },
...
        "api.WorkspaceKindEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/workspacekinds.WorkspaceKind"
                }
            }
        },
        "api.WorkspaceKindListEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspacekinds.WorkspaceKind"
                    }
                }
            }
        },
...
        "field.ErrorType": {
            "type": "string",
            "enum": [
                "FieldValueNotFound",
                "FieldValueRequired",
                "FieldValueDuplicate",
                "FieldValueInvalid",
                "FieldValueNotSupported",
                "FieldValueForbidden",
                "FieldValueTooLong",
                "FieldValueTooMany",
                "InternalError",
                "FieldValueTypeInvalid"
            ],
            "x-enum-varnames": [
                "ErrorTypeNotFound",
                "ErrorTypeRequired",
                "ErrorTypeDuplicate",
                "ErrorTypeInvalid",
                "ErrorTypeNotSupported",
                "ErrorTypeForbidden",
                "ErrorTypeTooLong",
                "ErrorTypeTooMany",
                "ErrorTypeInternal",
                "ErrorTypeTypeInvalid"
            ]
        },
...
        "storageclasses.StorageClassListItem": {
            "type": "object",
            "required": [
                "canUse",
                "description",
                "displayName",
                "name"
            ],
            "properties": {
                "canUse": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "workspacekinds.ImageConfig": {
            "type": "object",
            "required": [
                "default",
                "values"
            ],
            "properties": {
                "default": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspacekinds.ImageConfigValue"
                    }
                }
            }
        },
        "workspacekinds.ImageConfigValue": {
            "type": "object",
            "required": [
                "description",
                "displayName",
                "hidden",
                "id",
                "labels"
            ],
            "properties": {
                "clusterMetrics": {
                    "$ref": "#/definitions/workspacekinds.clusterMetrics"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspacekinds.OptionLabel"
                    }
                },
                "redirect": {
                    "$ref": "#/definitions/workspacekinds.OptionRedirect"
                }
            }
        },
        "workspacekinds.ImageRef": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
...
        "workspacekinds.PodConfig": {
            "type": "object",
            "required": [
                "default",
                "values"
            ],
            "properties": {
                "default": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspacekinds.PodConfigValue"
                    }
                }
            }
        },
        "workspacekinds.PodConfigValue": {
            "type": "object",
            "required": [
                "description",
                "displayName",
                "hidden",
                "id",
                "labels"
            ],
            "properties": {
                "clusterMetrics": {
                    "$ref": "#/definitions/workspacekinds.clusterMetrics"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspacekinds.OptionLabel"
                    }
                },
                "redirect": {
                    "$ref": "#/definitions/workspacekinds.OptionRedirect"
                }
            }
        },
        "workspacekinds.PodMetadata": {
            "type": "object",
            "required": [
                "annotations",
                "labels"
            ],
            "properties": {
                "annotations": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "workspacekinds.PodTemplate": {
            "type": "object",
            "required": [
                "options",
                "podMetadata",
                "volumeMounts"
            ],
            "properties": {
                "options": {
                    "$ref": "#/definitions/workspacekinds.PodTemplateOptions"
                },
                "podMetadata": {
                    "$ref": "#/definitions/workspacekinds.PodMetadata"
                },
                "volumeMounts": {
                    "$ref": "#/definitions/workspacekinds.PodVolumeMounts"
                }
            }
        },
        "workspacekinds.PodTemplateOptions": {
            "type": "object",
            "required": [
                "imageConfig",
                "podConfig"
            ],
            "properties": {
                "imageConfig": {
                    "$ref": "#/definitions/workspacekinds.ImageConfig"
                },
                "podConfig": {
                    "$ref": "#/definitions/workspacekinds.PodConfig"
                }
            }
        },
        "workspacekinds.PodVolumeMounts": {
            "type": "object",
            "required": [
                "home"
            ],
            "properties": {
                "home": {
                    "type": "string"
                }
            }
        },
...
        "workspacekinds.WorkspaceKind": {
            "type": "object",
            "required": [
                "deprecated",
                "deprecationMessage",
                "description",
                "displayName",
                "hidden",
                "icon",
                "logo",
                "name",
                "podTemplate"
            ],
            "properties": {
                "clusterMetrics": {
                    "$ref": "#/definitions/workspacekinds.clusterMetrics"
                },
                "deprecated": {
                    "type": "boolean"
                },
                "deprecationMessage": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "hidden": {
                    "type": "boolean"
                },
                "icon": {
                    "$ref": "#/definitions/workspacekinds.ImageRef"
                },
                "logo": {
                    "$ref": "#/definitions/workspacekinds.ImageRef"
                },
                "name": {
                    "type": "string"
                },
                "podTemplate": {
                    "$ref": "#/definitions/workspacekinds.PodTemplate"
                }
            }
        },
        "workspacekinds.clusterMetrics": {
            "type": "object",
            "required": [
                "workspacesCount"
            ],
            "properties": {
                "workspacesCount": {
                    "type": "integer"
                }
            }
        },
...
    }
}